requirements.txt
streamlit==0.88.0
pandas==1.3.3
praw==7.4.0
vaderSentiment==3.3.2
autocorrect==2.5.0
google-api-python-client==2.24.0
python-dotenv
display
os
